+++ src/migrations/m260604_000000_migrate_field_defaults.php
@@ -2,8 +2,8 @@
 namespace verbb\formie\migrations;
 
 use verbb\formie\Formie;
+use verbb\formie\helpers\MigrationHelper;
 
-use Craft;
 use craft\db\Migration;
 
 class m260604_000000_migrate_field_defaults extends Migration
@@ -17,7 +17,7 @@ class m260604_000000_migrate_field_defaults extends Migration
         $settings = $plugin->getSettings()->toArray();
         $settings = $plugin->getFormDefaults()->migrateLegacyFieldDefaults($settings);
 
-        Craft::$app->getPlugins()->savePluginSettings($plugin, $settings);
+        MigrationHelper::savePluginSettingsIfAllowed($plugin, $settings);
 
         return true;
     }